Add Suggestions Drawer
- Add DrawerModal component
- Removed Suggestion function from Slider.js
- Add DrawerModal to Booklist

The goal of the work thus far is to create a modal that displays
sugessted content based on the book current book in the carousel.
The issue I am currently having is that the suggestions are being
requested from an API, and passed down as a prop to the DrawerComponent, however the data is not loading in time to appear
in the drawer. Will continue to troubleshoot this issue.